+++ Task12/Program.cs
@@ -1,5 +1,5 @@
 ﻿Console.WriteLine("Введите число");
-long number = long.Parse(Console.ReadLine() ?? "0");
+int number = int.Parse(Console.ReadLine() ?? "0");
 int a = 99;
 int b = 999;
 
@@ -9,7 +9,7 @@ if (number>a)
    {
     number = number/10;
    }
-   long third = number%10;
+   int third = number%10;
 
 Console.Write(third);
 }
